requirements.txt
pyyaml==6.0.2
rdkit==2024.09.6
pandas==2.2.3
torch==2.6.0+cpu
torch_geometric==2.6.1
joblib==1.4.2
pydantic==2.10.6
sklearn==1.6.1
matplotlib==3.10.1
numpy==2.2.3
diskcache==5.6.3